# Refactor func args