--- a/api/image.go
+++ b/api/image.go
@@ -91,9 +91,6 @@ func BuildDockerImage(ctx context.Context, cli MeliAPiClient, dc *DockerContaine
 	if dockerFile == "" {
 		dockerFile = "Dockerfile"
 	}
-
-	// TODO: we should probably use the filepath stdlib module
-	// so that atleast it can guarantee us os agnotic'ness
 	formattedDockerComposePath := FormatComposePath(dc.DockerComposeFile)
 	if len(formattedDockerComposePath) == 0 {
 		// very unlikely to hit this situation, but
@@ -101,7 +98,7 @@ func BuildDockerImage(ctx context.Context, cli MeliAPiClient, dc *DockerContaine
 	}
 	pathToDockerFile := formattedDockerComposePath[0]
 	if pathToDockerFile != "docker-compose.yml" {
-		dockerFile = pathToDockerFile + "/" + dockerFile
+		dockerFile = filepath.Join(pathToDockerFile, dockerFile)
 	}
 
 	dockerFilePath, err := filepath.Abs(dockerFile)