`go test -v -race -cover -run=XXXX -bench=BenchmarkPullDockerImage ./...`
// new: 10171 ns/op, 10107 ns/op, 12604 ns/op
// old: 15388 ns/op, 15480 ns/op, 15884 ns/op

`go test -v -race -cover -run=XXXX -bench=BenchmarkBuildDockerImage ./...`
// new: 14773 ns/op, 14901 ns/op, 14622 ns/op
// old: 50439 ns/op, 47682 ns/op, 48038 ns/op

`go test -v -race -cover -run=XXXX -bench=BenchmarkCreateContainer ./... `
// new: 13233 ns/op, 12943 ns/op, 13016 ns/op
// old: 7282 ns/op, 6733 ns/op, 6764 ns/op

`go test -v -race -cover -run=XXXX -bench=BenchmarkContainerLogs ./...`
// new: 9266 ns/op, 9131 ns/op, 9134 ns/op
// old: 12715 ns/op, 12994 ns/op, 12405 ns/op