@komuW 2017-10-22
the benchmarks look good for the new implementation, or maybe it is coz I'm using `ioutil.Discard` in the benchmarks/tests :man_shrugging: